Add Service Coding

+++ coding-service/src/main/java/com/codecampus/coding/dto/response/SubmissionTestCaseResultDto.java
@@ -11,5 +11,5 @@ public record SubmissionTestCaseResultDto(
         String expectedOutput,
         String actualOutput,
         boolean passed,
-        long executionTimeMs) {
+        int executionTimeMs) {
 }
